+++ explore-with-me-stats/stats-dto/src/main/java/ru/practicum/dto/event/EventShortDto.java
@@ -17,5 +17,5 @@ public class EventShortDto {
     private UserShortDto initiator;
     private Boolean paid;
     private String title;
-    private int views;
+    private long views;
 }
